+++ src/index.js
@@ -1,7 +1,15 @@
-import invariant from './invariant'
+import assert from './assert'
 import isPlainObject from './isPlainObject'
+import parseDate from './parseDate'
+import formatDate from './formatDate'
+import getURLParams from './getURLParams'
+import accessCookie from './accessCookie'
 
 export default {
-  invariant,
-  isPlainObject
+  assert,
+  isPlainObject,
+  parseDate,
+  formatDate,
+  getURLParams,
+  accessCookie
 }