+++ CHANGELOG.md
@@ -28,3 +28,15 @@
 #### Features
 
 - Add module **getURLParams**.
+
+## 2.0.0 (2017-03-04)
+
+#### Features
+
+- Refactor and rename module **invariant** to **assert**.
+
+## 2.1.0 (2017-03-05)
+
+#### Features
+
+- Add module **accessCookie**.